Add support for SSH
Fork from the excellent upstream project https://github.com/juampe/docker-pi-hole-dot

This adds support for SSH which is not considered pure as it has a container doing more than a single job but this container is running on MikroTik RB5009 router. It is far easier to have SSH support for troubleshooting and tweaking in this kind of hosting environment.